fix : filter by trend, Refactored code
