+++ src/frontend/components/index.js
@@ -8,3 +8,5 @@ export { Filter } from "./filter/Filter";
 export { Post } from "./post/Post";
 export { Modal } from "./modal/Modal";
 export { UpdateProfile } from "./updateProfile/UpdateProfile";
+export { EditPost } from "./editPost/EditPost";
+export { PostComments } from "./postComments/PostComments";
