+++ src/frontend/features/bookmark/bookmarkSlice.js
@@ -1,5 +1,6 @@
 import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
 import axios from "axios";
+import toast from "react-hot-toast";
 
 const initialState = {
   data: [],
@@ -18,7 +19,6 @@ const getBookmarkPosts = createAsyncThunk(
       });
       return data;
     } catch (err) {
-      console.error(err);
       return rejectWithValue("Can't get bookmarked posts!");
     }
   }
@@ -39,7 +39,6 @@ const bookmarkPost = createAsyncThunk(
       );
       return data;
     } catch (err) {
-      console.error(err);
       return rejectWithValue("Can't save post!");
     }
   }
@@ -59,7 +58,6 @@ const removeBookmarkPost = createAsyncThunk(
       );
       return data;
     } catch (err) {
-      console.error(err);
       return rejectWithValue("Can't save post!");
     }
   }
@@ -80,30 +78,31 @@ const boomarkSlice = createSlice({
     [getBookmarkPosts.rejected]: (state, { payload }) => {
       state.isLoading = false;
       state.error = payload;
+      toast.error(state.error);
     },
     [bookmarkPost.pending]: (state) => {
       state.isLoading = true;
     },
     [bookmarkPost.fulfilled]: (state, { payload }) => {
-      console.log("bookmark : ", payload);
       state.isLoading = false;
       state.data = payload.bookmarks;
     },
     [bookmarkPost.rejected]: (state, { payload }) => {
       state.isLoading = false;
       state.error = payload;
+      toast.error(state.error);
     },
     [removeBookmarkPost.pending]: (state) => {
       state.isLoading = true;
     },
     [removeBookmarkPost.fulfilled]: (state, { payload }) => {
-      console.log("bookmark : ", payload);
       state.isLoading = false;
       state.data = payload.bookmarks;
     },
     [removeBookmarkPost.rejected]: (state, { payload }) => {
       state.isLoading = false;
       state.error = payload;
+      toast.error(state.error);
     },
   },
 });
