+++ src/frontend/features/user/singleUserSlice.js
@@ -99,6 +99,7 @@ const singleUserSlice = createSlice({
       state.isLoading = false;
       state.error = payload;
       state.status = "rejected";
+      toast.error(state.error);
     },
     [getUserAllPostsById.pending]: (state) => {
       state.isLoading = true;
@@ -112,6 +113,7 @@ const singleUserSlice = createSlice({
       state.isLoading = false;
       state.error = payload;
       state.status = "rejected";
+      toast.error(state.error);
     },
     [editUser.pending]: (state) => {
       state.status = "loading";
