+++ src/frontend/utils/getUserFeed.js
@@ -1,6 +1,6 @@
 export function getUserFeed(posts, following, currentUser) {
-  const followingUsername = following.map((user) => user.username);
-  return posts.filter(
+  const followingUsername = following?.map((user) => user.username);
+  return posts?.filter(
     (post) =>
       followingUsername.includes(post.username) || post.username === currentUser
   );
